refactor(screensharing): move preview to a new component
